Checkpoint for fromHits work

+++ src/alpaka/plugin-PixelSeeding/alpaka/CAStructures.h
@@ -71,8 +71,10 @@ namespace caStructures {
   using SequentialContainerView = typename SequentialContainer::View;
 
   template <typename TrackerTraits>
+  // using PhiBinnerT =
+  //     cms::alpakatools::HistoContainer<int16_t, 256, -1, 8 * sizeof(int16_t), hindex_type, TrackerTraits::numberOfLayers>;
   using PhiBinnerT =
-      cms::alpakatools::HistoContainer<int16_t, 256, -1, 8 * sizeof(int16_t), hindex_type, TrackerTraits::numberOfLayers>;
+      cms::alpakatools::HistoContainer<int32_t, 256, -1, 8 * sizeof(int32_t), hindex_type, TrackerTraits::numberOfLayers>;
 
   template <typename TrackerTraits>
   using CellNeighborsT =
